pyproject.toml
[tool.poetry]
name = "icon-creator"
version = "0.1.0"
description = ""
authors = ["Mike Lu <mcumaple@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests-oauthlib = "^1.3.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
